requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
python-multipart==0.0.5
aiofiles==0.6.0
apscheduler==3.7.0
python-docx==1.1.0
sqlalchemy==2.0.23
databases==0.8.0
mysql-connector-python==8.2.0
aiomysql==0.2.0
